requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
altair==6.0.0
altgraph==0.17.5
anki==2.1.66
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-generator==1.10
async-lru==2.1.0
attrs==25.4.0
Authlib==1.6.6
azure-core==1.38.1
azure-storage-blob==12.28.0
babel==2.17.0
bcrypt==5.0.0
beartype==0.22.9
beautifulsoup4==4.14.3
black==26.1.0
bleach==6.3.0
blinker==1.9.0
blis==1.3.3
boto3==1.42.47
botocore==1.42.47
brotli==1.2.0
cached-property==2.0.1
cachetools==6.2.5
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
chardet==3.0.4
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
coverage==7.13.4
cryptography==46.0.3
cssselect2==0.8.0
customtkinter==5.2.1
cycler==0.12.1
cyclopts==4.5.1
cymem==2.0.13
darkdetect==0.8.0
datasets==4.5.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fabric==3.2.2
fakeredis==2.33.0
fastapi==0.128.0
fastapi-cli==0.0.21
fastapi-cloud-cli==0.11.0
fastar==0.8.0
fastjsonschema==2.21.2
fastmcp==2.14.4
filelock==3.20.1
Flask==3.1.2
fonttools==4.61.1
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
genanki-mnemocards==0.11.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-generativeai==0.8.6
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
googletrans==3.1.0a0
greenlet==3.3.1
grpcio==1.78.0
grpcio-status==1.71.2
gTTS==2.5.4
gunicorn==25.0.3
h11==0.16.0
h2==3.2.0
hf-xet==1.2.0
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
hyperframe==5.2.0
identify==2.6.16
idna==2.10
importlib_metadata==8.7.1
iniconfig==2.3.0
invoke==2.2.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
iso-639==0.4.5
isodate==0.7.2
isoduration==20.11.0
isort==7.0.0
itsdangerous==2.2.0
jaconv==0.4.1
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.1.0
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keyring==25.7.0
kiwisolver==1.4.9
lark==1.3.1
librt==0.8.0
llvmlite==0.46.0
lupa==2.6
lxml==6.0.2
Mako==1.3.10
Markdown==3.10.1
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
mistune==3.2.0
mnemocards==0.1.5
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
mypy==1.19.1
mypy_extensions==1.1.0
narwhals==2.16.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6
nltk==3.9.2
nodeenv==1.10.0
notebook==7.5.3
notebook_shim==0.2.4
notebooklm-mcp-server==0.1.15
numba==0.63.1
numpy==2.3.5
openai==2.16.0
openai-whisper==20250625
openapi-pydantic==0.5.1
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
oracledb==3.4.1
orjson==3.11.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
paramiko==4.0.0
parso==0.8.6
pathable==0.4.4
pathspec==1.0.4
pathvalidate==3.3.1
patsy==1.0.2
pdfminer.six==20251230
pdfplumber==0.11.9
pefile==2024.8.26
pillow==12.1.0
platformdirs==4.5.1
playwright==1.58.0
playwright-stealth==2.0.2
plotly==6.5.2
pluggy==1.6.0
polars==1.38.1
polars-runtime-32==1.38.1
pre_commit==4.5.1
preshed==3.0.12
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.16.6
pydyf==0.12.1
pyee==13.0.0
PyGetWindow==0.0.9
PyGithub==1.59.1
Pygments==2.19.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
PyJWT==2.10.1
pykakasi==2.3.0
pymssql==2.3.12
PyMySQL==1.1.2
PyNaCl==1.6.2
pyodbc==5.3.0
pyparsing==3.3.2
pypdf==6.6.2
PyPDF2==3.0.1
pypdfium2==5.4.0
pyperclip==1.11.0
pyphen==0.17.2
pypiwin32==223
PyRect==0.2.0
PySocks==1.7.1
pystache==0.6.8
pytesseract==0.3.13
pytest==9.0.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytokens==0.4.1
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.0
referencing==0.36.2
regex==2025.11.3
reportlab==4.4.10
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.1
rich-rst==1.3.2
rich-toolkit==0.19.4
rignore==0.7.6
riotwatcher==3.3.1
rpds-py==0.30.0
ruff==0.15.0
s3transfer==0.16.0
safetensors==0.7.0
schedule==1.2.2
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
selenium==4.40.0
Send2Trash==2.1.0
sentencepiece==0.2.1
sentry-sdk==2.52.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.2
sse-starlette==3.2.0
stack-data==0.6.3
starlette==0.50.0
statsmodels==0.14.6
sympy==1.14.0
tenacity==9.1.4
terminado==0.18.1
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tkcalendar==1.6.1
tokenizers==0.22.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==5.1.0
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
typer==0.23.0
typer-slim==0.23.0
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
virtualenv==20.36.1
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.6.0
weasel==0.4.3
weasyprint==68.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.5
wheel==0.46.3
widgetsnbextension==4.0.15
win10toast==0.9
win11toast==0.36.3
winrt-runtime==3.2.1
winrt-Windows.Data.Xml.Dom==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.Globalization==3.2.1
winrt-Windows.Graphics.Imaging==3.2.1
winrt-Windows.Media.Core==3.2.1
winrt-Windows.Media.Ocr==3.2.1
winrt-Windows.Media.Playback==3.2.1
winrt-Windows.Media.SpeechSynthesis==3.2.1
winrt-Windows.Storage==3.2.1
winrt-Windows.Storage.Streams==3.2.1
winrt-Windows.UI.Notifications==3.2.1
wrapt==1.17.3
wsproto==1.3.2
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
youtube-transcript-api==1.2.4
zipp==3.23.0
zopfli==0.4.0
